// apps/cms/src/payload-types.ts
export interface Author {
  id: number;
  author: string;
  /**
   * Unique identifier for the author
   */
  author_id: string;
  /**
   * GitHub/social username
   */
  username?: string | null;
  position?: string | null;
  /**
   * Company name (for external/guest authors)
   */
  company?: string | null;
  /**
   * Link to GitHub, Twitter, LinkedIn, etc.
   */
  author_url?: string | null;
  author_image_url?: (number | null) | Media;
  updatedAt: string;
  createdAt: string;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "media".
 */
export interface Media {
  id: number;
  alt?: string | null;
  caption?: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  } | null;
  prefix?: string | null;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
  sizes?: {
    thumbnail?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    square?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    small?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    medium?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    large?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    xlarge?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
    og?: {
      url?: string | null;
      width?: number | null;
      height?: number | null;
      mimeType?: string | null;
      filesize?: number | null;
      filename?: string | null;
    };
  };
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "categories".
 */
export interface Category {
  id: number;
  name: string;
  parent?: (number | null) | Category;
  breadcrumbs?:
  | {
    doc?: (number | null) | Category;
    url?: string | null;
    label?: string | null;
    id?: string | null;
  }[]
  | null;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Post {
  id: number;
  title: string;
  slug?: string | null;
  slugLock?: boolean | null;
  /**
   * Appears as subheading in the blog post preview.
   */
  description?: string | null;
  content: {
    root: {
      type: string;
      children: {
        type: string;
        version: number;
        [k: string]: unknown;
      }[];
      direction: ('ltr' | 'rtl') | null;
      format: 'left' | 'start' | 'center' | 'right' | 'end' | 'justify' | '';
      indent: number;
      version: number;
    };
    [k: string]: unknown;
  };
  /**
   * Will show up as the blog post cover. Required.
   */
  thumb?: (number | null) | Media;
  /**
   * Authors must be one or more. Required.
   */
  authors?: (number | Author)[] | null;
  /**
   * Select only one category. Required.
   */
  categories?: (number | Category)[] | null;
  readingTime?: number | null;
  /**
   * Tags can be one or more. Optional.
   */
  tags?: (number | Tag)[] | null;
  toc_depth?: number | null;
  /**
   * Select a launch week to show launch week summary at the bottom of the blog post. Optional.
   */
  launchweek?: ('6' | '7' | '8' | 'x' | 'ga' | '12' | '13' | '14' | '15') | null;
  meta?: {
    /**
     * Defaults to the title of the post, if not set.
     */
    title?: string | null;
    /**
     * Defaults to the "thumb" image, if not set.
     */
    image?: (number | null) | Media;
    /**
     * Defaults to the description of the post, if not set.
     */
    description?: string | null;
  };
  publishedAt?: string | null;
  /**
   * This date will determine the chronological order of the blog post. Required.
   */
  date?: string | null;
  updatedAt: string;
  createdAt: string;
  _status?: ('draft' | 'published') | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "tags".
 */
export interface Tag {
  id: number;
  name: string;
  updatedAt: string;
  createdAt: string;
}